etl/requirements.txt
# ETL Pipeline Dependencies

# MongoDB
pymongo>=4.6.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Environment
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Web scraping (optional, for data extraction)
# scrapy>=2.11.0
beautifulsoup4>=4.12.0
# selenium>=4.15.0  # Replaced with Beautiful Soup for better performance

# Machine learning (for recommendation algorithms)
scikit-learn>=1.3.0

# Type hints
typing-extensions>=4.8.0

# Progress bars
tqdm>=4.66.0
